feat: refactor UniversalLink component using typescript
