+++ packages/volto/.eslintignore
@@ -1,1 +1,2 @@
 /types/
+!types/react-router-hash-link.d.ts
