+++ packages/volto/package.json
@@ -183,7 +183,7 @@
     "@plone/scripts": "workspace:*",
     "@plone/volto-slate": "workspace:*",
     "@redux-devtools/extension": "^3.3.0",
-    "classnames": "2.2.6",
+    "classnames": "2.5.1",
     "connected-react-router": "6.8.0",
     "debug": "4.3.2",
     "decorate-component-with-props": "1.2.1",
