--- packages/volto/types/components/manage/UniversalLink/UniversalLink.d.ts
@@ -1,22 +1,0 @@
-export default UniversalLink;
-declare function UniversalLink({ href, item, openLinkInNewTab, download, children, className, title, ...props }: {
-    [x: string]: any;
-    href: any;
-    item?: any;
-    openLinkInNewTab: any;
-    download?: boolean;
-    children: any;
-    className?: any;
-    title?: any;
-}): import("react/jsx-runtime").JSX.Element;
-declare namespace UniversalLink {
-    namespace propTypes {
-        let href: any;
-        let openLinkInNewTab: any;
-        let download: any;
-        let className: any;
-        let title: any;
-        let item: any;
-        let children: any;
-    }
-}
