+++ packages/volto/types/react-router-hash-link.d.ts
@@ -1,0 +1,12 @@
+declare module 'react-router-hash-link' {
+  import { ComponentType } from 'react';
+  import { NavLinkProps } from 'react-router-dom';
+
+  interface HashLinkProps extends NavLinkProps {
+    smooth?: boolean;
+    scroll?: (element: HTMLElement) => void;
+  }
+
+  export const HashLink: ComponentType<HashLinkProps>;
+  export const NavHashLink: ComponentType<HashLinkProps>;
+}
